+++ tests/func/grammar/test_keywords.py
@@ -67,10 +67,10 @@ def test_keyword_matches_on_word_boundary():
 
     g = Grammar.from_string(grammar)
     parser = Parser(g)
-    with pytest.raises(ParseError, match='forid=10 t" => Expected: for'):
+    with pytest.raises(ParseError, match='expected: for'):
         # This *will* raise an error
         parser.parse('forid=10 to20')
-    with pytest.raises(ParseError, match='Expected: to'):
+    with pytest.raises(ParseError, match='expected: to'):
         # This *will* also raise an error
         parser.parse('for id=10 to20')