+++ CHANGELOG.md
@@ -18,11 +18,15 @@ backward incompatible changes will start to apply when the projects goes 1.0
 
 ## [Unreleased]
 
+### Changed
+- Improved error reporting.
+
 ### Added
 - Support for Python 3.13 and 3.14.
 
 ### Changed
 - Moved to [uv](https://docs.astral.sh/uv/) for dependency/project management.
+
 
 ## [0.18.0] (released: 2024-02-23)